LDEV-3115, LDEV-3215: Grouping for Floating activities. Improve grouping detection for plain Tool activities. Last LD access list front end partial implementation.

+++ lams_common/src/java/org/lamsfoundation/lams/learningdesign/dao/ILearningDesignDAO.java
@@ -94,5 +94,5 @@ public interface ILearningDesignDAO extends IBaseDAO{
 	 */
 	public List getLearningDesignTitlesByWorkspaceFolder(Integer workspaceFolderID);
 	
-	public List<LearningDesignAccess> getAccessByUser(Long userId);
+	public List<LearningDesignAccess> getAccessByUser(Integer userId);
 }
